requirements.txt
requests
Pillow
keyboard
pyperclip
keyring
arabic_reshaper
python-bidi
